test: split protocol and contract suites
